requirements.txt
customtkinter==5.2.2
numpy==2.1.0
opencv_python==4.10.0.84
opencv_python_headless==4.9.0.80
Pillow==10.4.0
PyGetWindow==0.0.9
pytesseract==0.3.10
ttkbootstrap==1.10.1
